+++ submission/submission-webapp/src/main/kotlin/ac/uk/ebi/biostd/data/web/SubmissionDraftResource.kt
@@ -23,11 +23,13 @@ import org.springframework.web.bind.annotation.RequestBody
 import org.springframework.web.bind.annotation.RequestMapping
 import org.springframework.web.bind.annotation.ResponseBody
 import org.springframework.web.bind.annotation.RestController
+import springfox.documentation.annotations.ApiIgnore
 
 @RestController
 @RequestMapping(value = ["submissions/drafts"], produces = [APPLICATION_JSON_VALUE])
 @PreAuthorize("isAuthenticated()")
 @Api(tags = ["Submission Drafts"])
+@ApiIgnore
 internal class SubmissionDraftResource(private val subDraftService: SubmissionDraftService) {
     @GetMapping
     @ResponseBody
