+++ submission/submission-webapp/src/main/kotlin/ac/uk/ebi/biostd/groups/web/GroupsResource.kt
@@ -10,10 +10,12 @@ import org.springframework.security.access.prepost.PreAuthorize
 import org.springframework.stereotype.Controller
 import org.springframework.web.bind.annotation.GetMapping
 import org.springframework.web.bind.annotation.ResponseBody
+import springfox.documentation.annotations.ApiIgnore
 
 @Controller
 @PreAuthorize("isAuthenticated()")
 @Api(tags = ["Groups"])
+@ApiIgnore
 class GroupsResource {
     @GetMapping("/groups")
     @ResponseBody
