+++ submission/submission-webapp/src/main/kotlin/ac/uk/ebi/biostd/submission/web/resources/MultipartSubmitResource.kt
@@ -38,11 +38,13 @@ import org.springframework.web.bind.annotation.RequestParam
 import org.springframework.web.bind.annotation.ResponseBody
 import org.springframework.web.bind.annotation.RestController
 import org.springframework.web.multipart.MultipartFile
+import springfox.documentation.annotations.ApiIgnore
 
 @RestController
 @RequestMapping("/submissions")
 @PreAuthorize("isAuthenticated()")
 @Api(tags = ["Submissions"])
+@ApiIgnore
 @Suppress("LongParameterList")
 class MultipartSubmitResource(
     private val submitWebHandler: SubmitWebHandler,
