+++ submission/submission-webapp/src/main/kotlin/ac/uk/ebi/biostd/submission/web/resources/ProjectResource.kt
@@ -13,11 +13,13 @@ import org.springframework.web.bind.annotation.GetMapping
 import org.springframework.web.bind.annotation.RequestMapping
 import org.springframework.web.bind.annotation.ResponseBody
 import org.springframework.web.bind.annotation.RestController
+import springfox.documentation.annotations.ApiIgnore
 
 @RestController
 @RequestMapping("/projects")
 @PreAuthorize("isAuthenticated()")
 @Api(tags = ["Projects"])
+@ApiIgnore
 class ProjectResource(private val projectService: ProjectService) {
     @GetMapping
     @ResponseBody
